changed species from double to int

+++ libsrc/libsph/push_and_pull_u.c
@@ -82,7 +82,7 @@ int Push_and_Pull_U(PARTICLES *particles,PARS *pars,PARTICLES *x_particles)
     goto RETURN;
   }
   tag38  = 38*(MAX_TAGS)+rank;
-  if ( (MPI_Send(particles[0].species,nparticles,MPI_DOUBLE,cluster_size,tag38,MPI_COMM_WORLD)) != MPI_SUCCESS)
+  if ( (MPI_Send(particles[0].species,nparticles,MPI_INT,cluster_size,tag38,MPI_COMM_WORLD)) != MPI_SUCCESS)
   {
     printf(" %s(): unable to send tag38 packet \n",fname);
     exit_status = EXIT_FAILURE;
